Split Claude protocol and GitHub CLI modules
