+++ src/shared/claude/protocol/agent.ts
@@ -1,0 +1,39 @@
+/**
+ * Current task information for an agent.
+ */
+export interface AgentCurrentTask {
+  id: string;
+  title: string;
+  state: string;
+  branchName?: string | null;
+  prUrl?: string | null;
+}
+
+/**
+ * Assigned task information for an agent.
+ */
+export interface AgentAssignedTask {
+  id: string;
+  title: string;
+  state: string;
+}
+
+/**
+ * Agent metadata from WebSocket.
+ */
+export interface AgentMetadata {
+  id: string;
+  type: string;
+  executionState: string;
+  desiredExecutionState: string;
+  worktreePath: string | null;
+  /** Database session ID */
+  dbSessionId: string | null;
+  tmuxSessionName: string | null;
+  cliProcessId: string | null;
+  cliProcessStatus: string | null;
+  isHealthy: boolean;
+  minutesSinceHeartbeat: number;
+  currentTask?: AgentCurrentTask | null;
+  assignedTasks?: AgentAssignedTask[];
+}
