+++ src/shared/claude/protocol/index.ts
@@ -1,0 +1,12 @@
+export * from './agent';
+export * from './constants';
+export * from './content';
+export * from './guards';
+export * from './interaction';
+export * from './messages';
+export * from './models';
+export * from './queued';
+export * from './result-dedup';
+export * from './session';
+export * from './state-machine';
+export * from './websocket';
